feat(comments): доработан сервис комментариев, добавлено простое сидирование бд

+++ project/apps/comment/src/app/comment/comment.service.ts
@@ -1,10 +1,14 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { Comment } from '@project/contracts';
+import {
+  Injectable,
+  NotFoundException,
+  BadRequestException,
+} from '@nestjs/common';
+import { Comment, AccessTokenPayload } from '@project/contracts';
 import { CommentEntity } from './entity';
 import { Repository } from './service';
 import { CreateCommentDto } from './dto';
 import { Exception } from '../constants';
-import { PostQuery } from './validations';
+import { CommentQuery } from './validations';
 
 @Injectable()
 export class CommentService {
@@ -13,13 +17,16 @@ export class CommentService {
   /**
    * Создание комментария
    * @param payload Объект DTO
+   * @param user Данные access-токена
    * @returns Детали созданного комментария
    */
-  async create(payload: CreateCommentDto): Promise<Comment> {
+  async create(
+    payload: CreateCommentDto,
+    user: AccessTokenPayload
+  ): Promise<Comment> {
     const entity = new CommentEntity({
       ...payload,
-      // TODO: идентификатор авторизованного пользователя
-      author: '833a6872-29dd-4869-af2e-7df28a82aa6c',
+      author: user.id,
     });
     return this.repository.create(entity);
   }
@@ -27,7 +34,7 @@ export class CommentService {
   /**
    * Получение списка комментариев
    */
-  async getList(taskId: number, query: PostQuery): Promise<Comment[]> {
+  async getList(taskId: number, query: CommentQuery): Promise<Comment[]> {
     return await this.repository.getList(taskId, query);
   }
 
@@ -46,17 +53,22 @@ export class CommentService {
   /**
    * Удаление существующего комментария
    * @param commentId Идентификатор задачи
+   * @param user Данные access-токена
    */
-  async deleteItem(commentId: number): Promise<void> {
-    await this.findById(commentId);
+  async deleteItem(commentId: number, user: AccessTokenPayload): Promise<void> {
+    const record = await this.findById(commentId);
+    if (record.author !== user.id) {
+      throw new BadRequestException(Exception.BadRequest);
+    }
     await this.repository.delete(commentId);
   }
 
   /**
    * Удаление всех комментариев в разрезе определенной задачи
    * @param taskId Идентификатор задачи
+   * @param user Данные access-токена
    */
-  async deleteList(taskId: number): Promise<void> {
-    await this.repository.deleteCommentsList(taskId);
+  async deleteList(taskId: number, user: AccessTokenPayload): Promise<void> {
+    await this.repository.deleteCommentsList(taskId, user.id);
   }
 }
